feat(transport): decouple VFS gRPC from PyO3 + wire into nexusd-cluster
- Extract AuthProvider trait + ApiKeyAuth/NoAuth into services::auth
- Make transport::grpc.rs pure Rust (zero PyO3): auth via AuthProvider,
  Initialize/Call stubbed as Unimplemented
- Move PyBridge + Initialize + Call handlers to
  transport::python::grpc_bridge (Legacy* prefix, delete target)
- grpc module always compiled (no #[cfg(feature = "python")])
- Wire VFS gRPC into nexusd-cluster via shared port: cluster binary
  builds tonic::service::Routes, passes through ZoneManager to
  RaftGrpcServer::with_extra_services (type-erased, no circular dep)
- build_vfs_routes() returns Routes (VfsServiceImpl stays pub(crate))
- spawn() reuses build_vfs_routes() (DRY)
