+++ rust/profiles/cluster/Cargo.toml
@@ -23,7 +23,13 @@ contracts = { workspace = true }
 # HTTP stack so the binary stays at the sudowork target size.
 kernel = { workspace = true }
 backends = { workspace = true, default-features = false, features = ["driver-path-local"] }
+# VFS gRPC server — pure Rust content I/O on a separate port.
+# No python feature — cluster binary has zero PyO3.
+transport = { workspace = true }
+# Auth provider for VFS gRPC (NoAuth — mTLS is the boundary).
+services = { workspace = true }
 
+tonic = { workspace = true }
 anyhow = "1"
 clap = { version = "4", features = ["derive", "env"] }
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "signal", "time"] }
