+++ rust/raft/tests/test_async_bootstrap.rs
@@ -67,6 +67,7 @@ async fn make_node(node_id: u64, dir: &std::path::Path) -> (std::sync::Arc<ZoneM
         &bind_str,
         None,
         Some(format!("http://{bind_str}")),
+        None,
     )
     .expect("ZoneManager");
     (zm, bind_str)
