+++ rust/services/src/lib.rs
@@ -40,6 +40,11 @@
 //!                          +--- backends     (peer; never crosses to services)
 //! ```
 
+// Rust-native authentication providers (ApiKeyAuth / NoAuth / JwtAuth).
+// Always compiled — no feature gate. The transport tier consumes
+// `Arc<dyn AuthProvider>` to resolve bearer tokens without PyO3.
+pub mod auth;
+
 // AcpService — subprocess + ACP-over-stdio host for
 // `AgentKind::UNMANAGED` agents (claude / codex / …).  Currently
 // pyo3-laden internally, so the per-service gate also requires the
