+++ rust/transport/Cargo.toml
@@ -23,6 +23,9 @@ contracts = { workspace = true }
 # (mimalloc allocator, connectors HTTP, py-hook adapters) off in this
 # rlib; the cdylib pulls them via its own kernel dep.
 kernel = { workspace = true }
+# `services::auth::AuthProvider` trait consumed by grpc.rs for auth
+# resolution. No feature flags — auth module is always compiled.
+services = { workspace = true }
 # `lib::transport_primitives` (TLS / pool / addressing / TOFU /
 # PeerBlobClient trait) lives behind lib's `transport` feature.
 lib = { workspace = true, features = ["transport"] }
